requirements.txt
tweepy
requests
bs4
python-dotenv